chore(tech): rename cypress command
