+++ e2e-tests/cypress/e2e/journeys/quote/check-your-answers/check-your-answers-multi-policy.spec.js
@@ -35,7 +35,7 @@ context('Check your answers page (multiple policy) - as an exporter, I want to r
 
   before(() => {
     cy.login();
-    cy.submitQuoteAnswersHappyPathMultiPolicy();
+    cy.submitQuoteAnswersHappyPathMultiplePolicy();
     cy.url().should('include', ROUTES.QUOTE.CHECK_YOUR_ANSWERS);
   });
