+++ e2e-tests/cypress/e2e/journeys/quote/your-quote/your-quote-multi-policy.spec.js
@@ -33,7 +33,7 @@ context('Get a quote/your quote page (multiple policy) - as an exporter, I want 
   before(() => {
     cy.login();
 
-    cy.submitQuoteAnswersHappyPathMultiPolicy();
+    cy.submitQuoteAnswersHappyPathMultiplePolicy();
     submitButton().click();
 
     cy.url().should('include', ROUTES.QUOTE.YOUR_QUOTE);
