+++ e2e-tests/cypress/support/quote/submit-answers-happy-path-multiple-policy.js
